feat: add model submodule
